requirements/dev.txt
-r common.txt
astroid==1.3.6
logilab-common==1.0.2
pep8==1.6.2
pylint==1.4.4
pylint-django==0.6.1
pylint-plugin-utils==0.2.3
six==1.9.0
